Add gps init Refactor user button Refactor logging levels Add condition to LoRa
Added gps init function to make gps.py modular so it will be easier to implement getting the location.

User button reboot functionality is initialized at the beginning of main so it can be used to reboot if the sd card failed to initialize. Configuration functionality is only enabled later on when it is appropriate.

Changed some of the level of logging messages because they belong to debug rather than info.

LoRa does not send averages if none of the PM sensors are initialized.

Confirmed, that transistor circuitry works as intended on expansion board v4.

Discovered an interesting behaviour with configurations using expansion board v4 - takes couple of seconds to send configurations - sometimes completely fails.

+++ lib/LoRa_thread.py
@@ -22,47 +22,50 @@ def lora_thread(thread_name, logger, lora, lora_socket):
     :type timeout: int
     """
 
-    if lora_lock.locked():
-        logger.info("Waiting for other lora thread to finish")
-    with lora_lock:
-        logger.info("Thread: {} started".format(thread_name))
+    # Only send data over LoRa if at least one PM sensor is enabled
+    if config.get_config(s.PM1) != "OFF" or config.get_config(s.PM2) != "OFF":
 
-        try:
-            if lora.has_joined():
-                logger.info("LoRa connected")
-            else:
-                raise Exception("LoRa is not connected")
+        if lora_lock.locked():
+            logger.debug("Waiting for other lora thread to finish")
+        with lora_lock:
+            logger.debug("Thread: {} started".format(thread_name))
 
-            log_file_name = s.lora_file
+            try:
+                if lora.has_joined():
+                    logger.info("LoRa connected")
+                else:
+                    raise Exception("LoRa is not connected")
 
-            """Set the structure of the bytes to send over lora according to which sensors are defined. long_struct is 
-            used for temp+2PM, while short_struct is used for temp+1PM. Synchronized with back-end according to version 
-            number"""
-            structure = s.lora_long_struct
-            if config.get_config(s.PM1) == "OFF" or config.get_config(s.PM2) == "OFF":
-                structure = s.lora_short_struct
+                log_file_name = s.lora_file
 
-            if log_file_name not in os.listdir(s.lora_path[:-1]):  # Strip '/' from the end of path
-                raise Exception('Thread: {} - {} does not exist'.format(thread_name, log_file_name))
-            else:
-                with open(s.lora_path + log_file_name, 'r') as f:
+                """Set the structure of the bytes to send over lora according to which sensors are defined. long_struct is 
+                used for temp+2PM, while short_struct is used for temp+1PM. Synchronized with back-end according to version 
+                number"""
+                structure = s.lora_long_struct
+                if config.get_config(s.PM1) == "OFF" or config.get_config(s.PM2) == "OFF":
+                    structure = s.lora_short_struct
 
-                    # read all lines from lora.csv.tosend
-                    lines = f.readlines()
-                    for line in lines:
-                        stripped_line = line[:-1]  # strip /n
-                        split_line_lst = stripped_line.split(',')  # split line to a list of values
-                        int_line = list(map(int, split_line_lst))  # cast str list to int list
-                        logger.debug("Sending over lora: " + str(int_line))
-                        payload = struct.pack(structure, *int_line)  # define payload with given structure and list of averages
+                if log_file_name not in os.listdir(s.lora_path[:-1]):  # Strip '/' from the end of path
+                    raise Exception('Thread: {} - {} does not exist'.format(thread_name, log_file_name))
+                else:
+                    with open(s.lora_path + log_file_name, 'r') as f:
 
-                    lora_socket.send(payload)  # send payload to the connected socket
-                    logger.info("Thread: {} sent payload".format(thread_name))
-                    logger.info("Thread: {} removing file: {}".format(thread_name, log_file_name))
-                    os.remove(s.lora_path + log_file_name)
+                        # read all lines from lora.csv.tosend
+                        lines = f.readlines()
+                        for line in lines:
+                            stripped_line = line[:-1]  # strip /n
+                            split_line_lst = stripped_line.split(',')  # split line to a list of values
+                            int_line = list(map(int, split_line_lst))  # cast str list to int list
+                            logger.debug("Sending over lora: " + str(int_line))
+                            payload = struct.pack(structure, *int_line)  # define payload with given structure and list of averages
 
-        except Exception as e:
-            logger.exception("Sending averages over LoRaWAN failed")
-            blink_led(colour=0x770000, delay=0.5, count=1)
-        finally:
-            logger.info("Thread: {} finished".format(thread_name))
+                        lora_socket.send(payload)  # send payload to the connected socket
+                        logger.debug("Thread: {} sent payload".format(thread_name))
+                        logger.debug("Thread: {} removing file: {}".format(thread_name, log_file_name))
+                        os.remove(s.lora_path + log_file_name)
+
+            except Exception as e:
+                logger.exception("Sending averages over LoRaWAN failed")
+                blink_led(colour=0x770000, delay=0.5, count=1)
+            finally:
+                logger.debug("Thread: {} finished".format(thread_name))
